Chess stats added
Added basic statistics to user profile screen

+++ src/main/java/com/thombs/ChessWeb/Models/ChessGameDAO.java
@@ -6,5 +6,5 @@ public interface ChessGameDAO {
 	public ChessGame findGameByID(int id);
 	public ChessGame saveChessGame(ChessGame game);
 	public List<ChessGame> getAllChessGames();
-	public List<ChessGame> getChessGamesByUser(int userID);
+	public List<ChessGame> getChessGamesByUser(long userID);
 }
