+++ src/main/java/com/thombs/ChessWeb/Models/ChessGameService.java
@@ -6,5 +6,5 @@ public interface ChessGameService {
 	public ChessGame getChessGameByID(int id);
 	public List<ChessGame> getAllChessGames();
 	public ChessGame saveChessGame(ChessGame game);
-	public List<ChessGame> getChessGamesByUser(int userID);
+	public List<ChessGame> getChessGamesByUser(long userID);
 }
